Fix bug where CertValidity from EEprofile not used in CMP requests.
Fixed a bug where if the user specified a Certificate Validity
Start/End Time in the End User profile this value would be ignored
for CMP RA mode requests which create a new end entity based on that
EE profile.

The certificate validity information must be copied from the EE
profile into the EndEntityInformation object that is created when a
CMP message is received.

Note that this is still incomplete because the "modifiable" attribute
of the Certificate Validity from the EE profile is ignored.